#161 Moved some of gtfs_utils to 2 new files
