+++ gtfs/gtfs_utils/gtfs_utils/constants.py
@@ -1,0 +1,9 @@
+"""This file contains constants for paths and URLs related to GTFS"""
+
+MOT_FTP = 'gtfs.mot.gov.il'
+GTFS_FILE_NAME = 'israel-public-transportation.zip'
+LOCAL_ZIP_PATH = 'data/sample/gtfs.zip'
+TEMP_LOCAL_PATH = 'data/sample/gtfs_temp.zip'
+TARIFF_FILE_NAME = 'Tariff.zip'
+TARIFF_TXT_NAME = 'Tariff.txt'
+TARIFF_TO_REFORM_ZONE = 'StationToReformZone.txt'
