+++ gtfs/gtfs_utils/gtfs_utils/download_daily_ftp_files.py
@@ -1,7 +1,7 @@
 import datetime
 import os
 import re
-import gtfs_utils as gu
+from . import ftp
 
 CONF = {
     'local_dir': 'C:\\dev\\ds\\open-bus-explore\\data\\archive\\',
@@ -11,7 +11,7 @@ CONF = {
 
 def get_files(local_dir=CONF['local_dir'], size_limit=CONF['size_limit']):
     files = [parsed['file_name'] for parsed in
-             (re.search(gu.RE_FTP_LINE, line) for line in gu.get_ftp_dir())
+             (re.search(ftp.RE_FTP_LINE, line) for line in ftp.get_ftp_dir())
              if parsed['size'] != '<DIR>'
              and int(parsed['size']) <= size_limit]
 
@@ -23,7 +23,7 @@ def get_files(local_dir=CONF['local_dir'], size_limit=CONF['size_limit']):
         pass
 
     for file in files:
-        gu.get_ftp_file(file_name=file, local_path=daily_dir + file, force=True)
+        ftp.get_ftp_file(file_name=file, local_path=daily_dir + file, force=True)
 
 
 def main():
